+++ crates/ws/Cargo.toml
@@ -38,6 +38,7 @@ futures-rustls = "0.24"
 glommio = "0.8"
 hashbrown = { version = "0.14", features = ["serde"] }
 httparse = "1"
+indexmap = "2"
 log = "0.4"
 metrics = { version = "0.21", optional = true }
 metrics-util = { version = "0.15", optional = true }
@@ -50,6 +51,7 @@ rustls-pemfile = "1"
 serde = { version = "1", features = ["derive"] }
 signal-hook = { version = "0.3" }
 slab = "0.4"
+slotmap = "1"
 socket2 = { version = "0.5", features = ["all"] }
 tungstenite = "0.20"
